Conflicts:
	Raven.Database/DocumentDatabase.cs

+++ Raven.Tests/RavenTest.cs
@@ -33,7 +33,7 @@ namespace Raven.Tests
 
 		private string path;
 
-		public EmbeddableDocumentStore NewDocumentStore(string storageType = "munin", bool inMemory = true, int? allocatedMemory = null)
+		public EmbeddableDocumentStore NewDocumentStore(string storageType = "munin", bool inMemory = true, int? allocatedMemory = null, bool deleteExisting = true)
 		{
 			path = Path.GetDirectoryName(Assembly.GetAssembly(typeof(DocumentStoreServerTests)).CodeBase);
 			path = Path.Combine(path, "TestDb").Substring(6);
@@ -54,9 +54,10 @@ namespace Raven.Tests
 				ModifyStore(documentStore);
 				ModifyConfiguration(documentStore.Configuration);
 
-				if (documentStore.Configuration.RunInMemory == false)
+			if (documentStore.Configuration.RunInMemory == false && deleteExisting)
 					IOExtensions.DeleteDirectory(path);
 				documentStore.Initialize();
+
 
 				CreateDefaultIndexes(documentStore);
 
@@ -128,10 +129,6 @@ namespace Raven.Tests
 				Thread.Sleep(25);
 		}
 
-		protected virtual void ConfigureServer(RavenConfiguration configuration)
-		{
-		}
-
 		protected void WaitForUserToContinueTheTest()
 		{
 			if (Debugger.IsAttached == false)
@@ -178,11 +175,11 @@ namespace Raven.Tests
 			{
 				using (var documentStore = new DocumentStore
 											{
+												Url = "http://localhost:" + port,
 												Conventions =
 													{
 														FailoverBehavior = FailoverBehavior.FailImmediately
 													},
-												Url = "http://localhost:" + port
 											}.Initialize())
 				{
 					CreateDefaultIndexes(documentStore);
